commit de prueba del tp

+++ Lima_Natanael_3907_TP4/src/main/java/ar/edu/unju/fi/repository/IClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ar.edu.unju.fi.tp4.model.Cliente;
 
-public interface IClienteRepository extends CrudRepository<Cliente, Integer>{
+public interface IClienteRepository extends CrudRepository<Cliente, Long>{
 	
 }
